Ensuring URL state for team page tabs
* navigation between tabs is retained in the URL
* enabling better UX when using the back button
* enabling links back to certain tabs, such as the services tab after viewing the service detail page
